Fixed layer re-ordering feature
People can type in what layer number they want to move a layer to in order to change the visual appearance. They cannot enter a number outside the range of the current number of layers, blank text, or 0. 
The key part of this was calling the move_child function on the LayerList scene itself.