Added visualization, fixed layering issues
Layers are now visualized. There are still some issues with the offset and setting location on the canvas, but the basic functionality is there.
Fixed layering method. Layers are now stored and accessed in an array rather than directly as the children of the canvas. This is because there were unexplainable issues when the layers were renamed.